Move data transformation to the shared area #53
